Renamed motorsimulation
